Bug #31603195: REGRESSION: CTE CRASHING IN FINDSINGLEITERATOROFTYPE
When we have a recursive CTE and instantiate the MaterializeIterator,
it needs a pointer to the child FollowTailIterator. However, in the
rare case that said iterator has been optimized away, we need to simply
turn off the recursion. There was code to check for this case in
CreateIteratorFromAccessPath (FindSingleIteratorOfType returning nullptr),
but it was incomplete and inadvertently dereferenced a nullptr access path
on the way. Fix by simply testing for a nullptr access path and then
return a nullptr iterator.

Change-Id: I40d5223c9e32037a79b59f79b8140f02650d6ccf