Bug#31216115: WHEN WHERE CONDITION REFERS TO SUBQUERY AND IS ALWAYS  	      FALSE, SERVER WILL CRASH.
Problem:
Removing a sub-query present as part of a condition which would evaluate
to "always false", results in server exit.
The subquery that gets removed is used by its alias and is present in
the projection list of a merged derived table. But the outer select
to which, this derived table gets merged into, does not have it as part
of its projection. So the existing test to see if the subquery being
removed is present in the select list of the query does not suffice.
Hence removing this subquery  makes the access by the next reference
invalid.

Solution:
Do not call clean-up-after-removal if the item is of type Item_view_ref
which are created only for expressions from merged derived table.

Approved By : Dag Wanvik <dag.wanvik@oracle.com>