Bug #31649037: REMOVE LEX->CURRENT_SELECT() DEPENDENCY FROM RANGE OPTIMIZER
The LEX structure has a member called "current_select" which is marked as
parser-only; however, a few things use it under execution and expect it to be
set. (There are also some uses under resolving and optimization.) This is
a bit wasteful, and leads to hard-to-diagnose bugs if it is not properly
kept up-to-date at every point. One of those is the range optimizer, which just
as well can accept the query block in question as a parameter.

Change-Id: I2c19e0278ea92e5eadf5ee70151da3d404792002