Bug#31253103: UNCLEAR SYNTAX FOR CHANGE MASTER TO MANAGED Bug#31514952: WL#12649: REPLACE PRIMARY/MASTER BY SOURCE AND SECONDARY/SLAVE BY REPLICA
Problem:
========
The CHANGE MASTER TO MANAGED=1 enables a built-in feature to automatically
re-establish an asynchronous replication connection to another sender
(a primary), in case the current connection gets interrupted due to the failure
of the current sender. But the name MANAGED does not precisely convey the
actual task accomplish by this feature.

The WL#12649 is using slave and master terminology which is sensitive term
and should be replaced.

Solution:
=========
Rename all usage of primary/master by source, and secondary/slave by replica.
Also rename MANAGED option of CHANGE MASTER command to
SOURCE_CONNECTION_AUTO_FAILOVER.

RB: 24521