Bug#31582758 SUPPRESS DUPLICATE KEY ERRORS ALSO FOR ANALYZE
If the statistics table is updated concurrently, there is a chance
that ANALYZE TABLE may fail with "Duplicate key error" if a record was
inserted in the interval between the check for the existence of the
record and the execution of the insert. This is very rare situation.
Hence we ignore the error and clear the DA, in this situation.

This bug relates to Bug#30769965. So, the test case for Bug#30769965
in main.information_schema_statistics is extended.

Change-Id: Ie3681315e9679854dad7e7213c7110367b107b8d