BUG#31036860 MTR DOES NOT CREATE TRACE FILES IF --DEBUG              OPTION IS SPECIFIED ON WINDOWS
Issue:

When the --debug option is specified to generate trace
files on Windows, MTR strips off the drive name from
the trace file path. This in-turn causes trace files
to be absent in case the MTR vardir is on a different
drive than the current one.

Fix:

Instead of using a unix-like path, MTR now uses the
complete path to debug trace files which includes the
drive name on Windows.

Change-Id: I09745a9d167f0553ee9d9e05e4084977d944218b