Bug#31525210 IPV6 SUPPORT CHANGES BEHAVIOR WITH MDNS .LOCAL HOSTNAME
Problem
=======

IPv6 have the notion of zone index which is used to indicate for which
network interface a link local address is valid.

It has been seen on a MacOS machine that a host name lookup resulted in
a link local IPv6 address, which Ndb can not handle since it can not
propagate the zone index (interface number) when connecting or binding a
socket.

Solution
========

When resolving hostnames Ndb will choose in order:

1) If there are some IPv4 interface: the first IPv4 address found.
2) If there are some IPv6 interface: the first IPv6 address without zone
   index.

Note 1
======

The preference to IPv4 addresses are choosen to ease upgrades from ndb
nodes not aware of IPv6.

Note 2
======

As for old ndb nodes having a hostname resolve to more than one IPv4
address, or now also more than one IPv6 address, may have strange
effects if not all nodes see the same address order.

Note 3
======

Ndb uses sockets with IPv6 addresses internally and depends on dual
stack support.  IPv4 mapped addresses will trigger use of IPv4 protocol
using the corresponding IPv4 address.

When using getaddrinfo() the flags AI_V4MAPPED and AI_ADDRCONFIG do not
work well together.

The IPv4 mapped address are treated as an IPv6 address (which it is)
and will be returned by getaddrinfo() in an IPv6 only environment even
if unusabled due to the lack of IPv4 interfaces.

Also the IPv4 mapped address will not be returned if there are only IPv4
interfaces by getaddrinfo() if AI_ADDRCONFIG is in use.

Note 4
======

Link-local or scoped IPv6 addresses are not supported.

One reason is that we currently do not pass around a full sockaddr_in6
which contains the scope id (aka zone index).  Which would be needed in
calls to connect() and bind().

Neither have we any code that print nor read zone index in text form.

And having the notion of an address that can refer to different hosts in
different contexts is new and may have further consequences for Ndb that
one need to consider.

Reviewed-by: John David Duncan <john.duncan@oracle.com>
Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>