Bug#30786762: FAILED ASSERTION FOR DATES WITH ZERO OR ABOVE TIMESTAMP LIMIT
This bug report unfortunately contained two independent
bugs. The first one concerns bounds checking for timestamp
literals and was fixed along with bug no 31239157.

This patch fixes the second bug, failed assertion for
timestamp literals with zero month and a time zone
displacement. We calculate the number of seconds since the
unix epoch in order to adjust time zone diplacement, and the
function doing this, sec_since_epoch64(), asserts that the
month be non-zero.

We fix this by adding a check for zero month, and for
consistency we disallow a day zero as well, for timestamp
literals with a time zone displacement, raising an
error. This check runs independent of sql_mode.

Change-Id: Idfea577c6528afa2d9bd5f30b2c3d8db9525bd8d