Bug#30988735: SELECT QUERY WITH ERROR RESULT WITH "NOT IN" WHERE_CLAUSE
Problem:
When range values specified in a predicate are not compatible
with the field's datatype that these values are compared with,
range optimizer rounds off the range values and assigns certain
flags so that it does not exclude rows that qualify for the
range because of rounding.
In the problematic query we have "id not in (-0.1, 0.1)". Range
optimizer will first round off both the values to integers as
the field "id" is of type integer. So the range would now become
"id not in (0,0)". To overcome the problem of missing out rows,
once this range is constructed, range optimizer compares the
rounded value to the original value specified. Based on this,
flags get set. For ex if the rounded off value is greater than
the original value which is the case when "-0.1" gets rounded off
to "0", for a "<" operator, it sets flags such that the it
stops fetching values before the key value "0" (HA_READ_BEFORE_KEY)
and similarly for all other combinations, flags are set - for ex
if the rounded off value is lesser than the original value and if
its a ">" operator then a different flag gets set so that row after
the key value "0" is fetched ("HA_READ_AFTER_KEY").

For a "NOT IN" predicate, range optimizer constructs ranges in
the following way:
"id not in (1,2)" will be divided into 3 intervals -
"id < 1", "1 < id < 2", "id > 2"
While constructing the second interval it fetches the max value
from the first interval and sets as the starting range for the
second interval. In this case it would be "1". However for the
second interval, it should not be reading rows having id = 1,
instead it should be starting after the value "1". So it sets
"NEAR_MIN" flag which translates to "HA_READ_AFTER_KEY".
On the same lines, for the problematic query we have "not in (0,0)",
which would translate to - the first interval as "id < 0", the
second interval would be "0 < id < 0" with "NEAR_MIN" flag set
which translates to "Start reading after 0" which is not correct.
Had we not rounded off it would have been "Start reading after '-0.1'"
which would have included a value "0" in the interval.
This flag setting will later lead to merging of range intervals
leading to wrong results.

Solution:
While constructing the intervals for not in operator after the
starting interval is constructed, we now check for rounding and
then set the flag accordingly.

Change-Id: I95d2465cb11187967a09c1172179401f283eeb11