Bug#31665109 TROUBLE STARTING NDBMTD WHEN USING SYSTEM HOSTNAME
The work for wl#13860 in match_hostname() uses IN6_IS_ADDR_LOOPBACK()
to test an address, but this macro only tests for the IPv6
loopback address, and not for the mapped representation of the
IPv4 loopback address. Fix by relplacing the macro call
with a more thorough test.

The particular wording of the error message "Connection done
from wrong host" (from MgmtSrvr.cpp line 4216) can be
considered the signature of this bug.

Also apply some remaining changes from wl#13860 code review.

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>