Bug #31617670 MCC/MCM IS NOT ABLE TO START CLUSTER ON WINDOWS AFTER THE IPV6 PUSH
Problem
-------

By default, an IPv6 socket created on Windows Vista and later only operates
over the IPv6 protocol.

Solution:
---------

In order to make an IPv6 socket into a dual-stack socket,
the setsockopt function must be called with the IPV6_V6ONLY socket option to set
this value to zero before the socket is bound to an IP address.
When the IPV6_V6ONLY socket option is set to zero, a socket created for the
AF_INET6 address family can be used to send and receive packets to and from an
IPv6 address or an IPv4 mapped address.

Patch also includes an unrelated IP address formatting fix.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>