Bug#31687295 MTR MAY USE A TRUNCATED EXECUTABLE PATH WITH DEBUGGERS
MTR checks if the binary path returned by a debugger exists and is
executable. In case the path is truncated and points to an existing
directory, it will proceed to use the path to analyze core dumps
and thereby fails to generate the stack trace. An additional check
whether the binary path is a plain file is done before calling the
debugger to avoid such a situation.

Change-Id: I090d1da3ee97f50c35ce825e04d10d08a4e68978