Bug#31016076 SELECT COUNT(*) ON LARGE TABLE HANG FOREVER EVEN AFTER CTRL+C
Issue
=====
When aborting a select count(\*) on a table with large amount of data the
system hangs.

Although the system appears to be hung the case is that the abortion of the
query is waiting for all the parallel reader worker threads to quit. So what
appears as a hang is essentially select count(*) working in the background.

Fix
===
The fix is to check for the trx being interrupted while traversing the
records, and not count on the callback to propogate the information to the
Parallel reader.

The patch consists of two parts:
a) The actual fix (search for trx_is_interrupted() function call)
b) refactoring of the code where we make sure that we're consistently
setting the error state in scan_ctx and in the global error state.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>
RB: 24802