BUG#31577633: Missing LCP completion call
Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>

In rare circumstances we could miss calling the check for LCP completion
thus leaving an LCP uncompleted and thus LCPs cannot proceed.

Added lots of debugging to find the issue and added 3 more places where
the the call to check for LCP completion can be useful. The move of the
call in tableCloseLab is likely the main contributor to hangs.