WL#14081: Improved handling of real-time break scans
Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>

Each time we move to a new row in Dblqh when scanning we decide whether
we should continue scanning or not. We base this decision on the job buffer
level. However this means that one complex query can create job buffer load
that protect itself from itself.

To avoid this kind of feedback we track the following things in
Dblqh at a detailed level.

1) Number of scan fragment and scan next requests from DBSPJ
2) Number of scan fragment and scan next requests from DBTC
3) Number of key lookups from DBSPJ
4) Number of key lookups from DBTC
5) Number of key lookups for write from DBTC

Using this information we first check if more than 99% of the
events arrive from DBSPJ, in this we decrease job buffer level
and thus increase the number of rows scanned per real-time
break.

We also check if key lookup events are much more frequent than scans,
if so we do the same increase of the rows scanned
per real-time break.

Similarly if the number of scanned rows is more than 200 times
larger than the number of key lookups (a write key lookup is
counted as 3), then we also increase the number of rows scanned
per real-time break since there are very few key lookups to give
higher priority.

The idea with checking number of scanned rows is to also catch
a number of scans that scans a lot of rows in a few scan events.

We still decrease the number of rows scanned in very high load
cases to ensure that short scans have some prio over long scans
although not at the same level as key lookups are prioritised over
scans.