WL #13459: Optimize hash table in hash join [patch 3/9, memory limit] Bug #31516149: IN-MEMORY HASH JOIN WILL ONLY USE TWO-THIRDS OF JOIN BUFFER [noclose]
When MEM_ROOT has a capacity limit and wants to allocate a new block
(1.5x as large as the previous one), it errors out if the desired size of the
block would take us over capacity. This means we'd have spare capacity
that would not be used; e.g. hash join would typically use 212 kB out of
256 kB. (The bug report claims 2/3 of the buffer, but that is not
representative for most sizes; in practice, it's usually higher, although
there do exist cases that are that bad.)

Since often a smaller block could still satisfy the allocation,
try allocating up to exactly the capacity, and see if that is good
enough.

We did already have unit tests for this, but they did not check that the
allocation actually succeeded, so update them.

Change-Id: Id04d7ef4e3432356a4c21bc33b95ef2b0b3af869