WL #13459: Optimize hash table in hash join [patch 9/9, switch hash maps] Bug #31516149: IN-MEMORY HASH JOIN WILL ONLY USE TWO-THIRDS OF JOIN BUFFER
Switch hash join into using robin_hood::unordered_hash_map, significantly
reducing both CPU and memory usage. This is now mostly just a drop-in switch
now, but it allows a few extra optimizations:

 - Since the new map supports hetereogenous key lookups (a feature from C++17),
   we can store the keys using ImmutableStringWithLength, but still query
   them using Key (our std::string_view equivalent). This allows more efficient
   length storage.
 - robin_hood:: ships with its own hash that is faster than XXHash64 for
   small keys, and probably still good enough for our uses, so we switch
   to it. (We keep XXHash64 for chunk splitting, since robin_hood::hash_bytes
   does not support variable seeds.)

Memory management is changed; we now actually tell the MEM_ROOT what its
capacity is. Combined with the previous capacity patches, this means it will
not actually go over the limit (as opposed to earlier, where it would allocate
a block of memory it couldn't actually use for anything). The rest of the code
expects that an insertion will always succeed, which is an assumption that's
fairly hard to untangle, so the very last row gets inserted into a special
“overflow” MEM_ROOT with only one row and no limit.

Instead of the hash map being on the MEM_ROOT, it is now allocated on the
regular heap, so that the old buckets can be freed when the hash map grows.
Every time we insert into it, we adjust the MEM_ROOT's capacity (the new hash map
supports querying for memory usage), so that if the resize causes us to grow
over the limit, we will immediately spill to disk on the next row.

Microbenchmarks compared to before the entire worklog (GCC 10.1, 64-bit,
Coffee Lake 3.2 GHz):

  BM_HashTableIteratorBuild            1083326 ->  678894 ns/iter [+59.6%]
  BM_HashTableIteratorProbe            1090914 ->  730319 ns/iter [+49.4%]
  BM_HashTableIteratorProbeSemiJoin     684753 ->  563218 ns/iter [+21.6%]

Change-Id: I066ff0e0db1532dd2a6f9303901a1de6ec89161a