Bug#31721534 MTR TEST CASE NDB_BINLOG_DDL_MULTI FAILS IN PB2
The coordinator uses the node unsubscribe events to take note of any
participants failing or shutting down during a schema distribution, so
that it can stop waiting for a reply from that participant server. But
if the coordinator recieves a node unsubscribe event before it has
registered the participants for an ongoing schema operation, it
wrongly mistakes the non existent participants as a completion of the
schema op distribution itself and sends a final ack. When the schema
op event eventually reaches the coordinator for actual processing, it
ends up failing the server.

This patch fixes that by making the coordinator to check for only the
schema operations it is waiting on, when a node unsubscribe event is
received. Added a new testcase to verify the new behaviour.

Change-Id: Ia01fea2247754ae0a802d29ccd707585c5ec6d5b