Implemented PXB-2254 (Redesign --lock-ddl-per-table)
https://jira.percona.com/browse/PXB-2254

The current implementation of --lock-ddl-per-table works as follow:

 - Creates a dedicated mysql connection for MDL and start a transaction.
 - Start to scan directories (databases) looking for .ibd files.
 - For each .ibd file it tries to transalate the space id recorded on
the file header to a table (or tables in case of shared tablespace).
 - If it finds a table associate with the .ibd (or tables) it executes a
SELECT * FROM table LIMIT 1 using the dedicated connection for MDL.
 - From this point forward, the table is locked and new DDL will be
blocked until the MDL connection commits.
 - With this, it has been assumed that if a MLOG_LOAD_INDEX event has
been generated (https://dev.mysql.com/worklog/task/?id=7277) it's safe
to continue since we have taken MDL on the table we are copying.

This design has a few flaws:

 1. FTS have their own tablespace, which is been copied without
MDL protection.
 2. Ongoing DDL that generate temporary table (#sql...) are copied
without MDL protection.
 3. When creating FTS for the first time, MySQL will have to create a
tmp table (Item 2) and FTS files (Item 1) are pointing to temp table.
 4. Table space IDs for FTS or regular tables will change if you
DROP and then CREATE the same FTS/TABLE. MDL lock query will fail to
find the corresponding table but file will exist, thus will be copied
without MDL protection.
 5. If a table is created after xtrabackup has gathered the list of
.ibd's to copy and before the backup has finished, this particular
table will be part of the backup as it will be contained within the
redo log. However, if a bulk index load operations have been performed
( MLOG_INDEX_LOAD has been written to the redo log), this change to the
data file will not be tracked and will be missed from the backup.

This commit changes to:

1. MDL is taken at the beginning of backup for all know tables at
that time, before we do the first scan of redo logs.
2. A new flag controls whatever MDL step has been completed or not. We
initiate this flag with false.
3. First run of redo log scan accepts parsing MLOG_INDEX_LOAD.
At this point is still safe to accept it since we have not started to
copy .ibd files and if MLOG_INDEX_LOAD events have happened right before
we start the backup, the first run will identify them.
4. After first scan of redo logs, mdl_taken changes to true and parsing
a MLOG_INDEX_LOAD causes the backup to abort.
5. Before we execute the MDL query for a specific entry from
INNODB_SYS_TABLES, we validate it against a regexp for a FTS file name.
if we are manipulating a FTS index, we skip MDL query as we will
eventually cross the FTS base table while iterating through the list of
innodb tables.
6. If for any reason we cannot execute the MDL lock query we do not
abort the backup. Since we change mdl_taken to true after redo log has
caught-up and before starting to copy .ibd's its safe to allow backup
to continue. We will only abort if a subsequent MLOG_INDEX_LOAD is
found by the redo follow thread during the backup.
7. We identify temporary tables (#sql-) and skip trying to lock them.
No reason to attempt to query it as it will always result in an error.
8. MDL query changed to not bring any row back. This is enough for
taking a MDL while saving the server from retrieving any data.