Bug#31667061: Inconsistent behavior of @@SQL_SELECT_LIMIT with prepared statements
The problem manifested here is that varying values of the variable
sql_select_limit are not effective for repeated executions of prepared
statements. Instead, the value of sql_select_limit as of when the
statement was prepared is used.

WL#9384 had to refactor this handling, because we had to separate
preparation and execution of prepared statements better. However,
it was decided to take the value of sql_select_limit at preparation
time and materialize it as a constant integer item, which was later
evaluated per execution.

The solution to the problem is to remove the constant integer item and
instead flag that the value of sql_select_limit must be evaluated
per execution. (We could have created an Item_get_user_var object for
this, however this means a longer code path and the idea was rejected).
The flag is called m_use_select_limit and is set for the outermost
query block of SELECT statements (the variable is not effective for
e.g UPDATE or DELETE statements).

Since we no longer have a need to distinguish "explicitly" created
items that represent row count limits, the variable explicit_limit
has been removed. However, we still use internally generated limit
objects with subqueries, and since we don't want to display these in
EXPLAIN statements, we added a new flag m_internal_limit to identify
these.

The property function SELECT_LEX::has_explicit_limit_or_order() that
depended on explicit_limit was deleted, since it is easy to replace
with other functions.

Some outdated logic and comments in SELECT_LEX::get_limit() and
SELECT_LEX::get_offset() has also been removed.

Also adding SELECT_LIMIT for SHOW commands, as well as a test case.

Reviewed by: Chaithra Gopalareddy <chaithra.gopalareddy@oracle.com>