BUG#31726476: testIOBasic -case d problem
Approved-by: Mauritz Sundell <mauritz.sundell@oracle.com>

testOIBasic -case d
reports error when a row is reported twice in a range scan. This is however
expected behaviour as NDB scans one row at a time using the range order.
If a row is updated it will move in the range index, if the row is first
found with a low range order and then updated to move it forward it can be
found again and even more than two times. So removed error reporting for this
behaviour.