BUG#31782578: Abort arrives before LQHKEYREF
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

If DBTC has a timeout it can send ABORT while a node is still processing
an LQHKEYREQ, this LQHKEYREQ can report LQHKEYREF to a node that already
received ABORT. In this case we need to ignore the LQHKEYREF since the
operation might have been reused for something new. If this something is
a COPY fragment operation this wasn't taken care of.