Bug #31762806: UNNECCESSARY TEMPORARY TABLE BEFORE COUNT(DISTINCT)
In certain queries with DISTINCT aggregates (which are generally solved by sorting
before aggregation), we use a temporary table instead of streaming, because we
mistakenly think that the temporary table does deduplication. This is because
we try to replicate create_tmp_table()'s logic for when we get a unique index,
but fail to take into account param->allow_group_via_temp_table. Fix by
checking for the unique index instead, which is more robust and takes even more
logic out of QEP_TAB.

Change-Id: If4918144361668affaa9eea426976c74da92547f