Bug#31301101 ASSERTION ".CONV_FIELD->HAS_CHARSET"
By running an ALTER TABLE <> CONVERT TO CHARACTER SET command, the
character set of every CHAR, VARCHAR and TEXT columns is updated to the
new value. Internally, an ARRAY column that is used for a multi-valued
index uses a hidden CHAR column which can only be
my_charset_utf8mb4_0900_bin
or binary.
The fix is to avoid setting the hidden column's charset to that of the
table, similar to what's done for BLOB columns.

Change-Id: I9eb455910b3d670953e7a7518639ffd4b44fe100