Bug #31397578: MAKE TESTS RESILIENT TO ORDER CHANGES [noclose]
Yet more test changes to prepare for the hypergraph optimizer.

In particular, we disable large parts of the perfschema suite for the
hypergraph optimizer, since they by nature are very dependent on what
plans are chosen. (They would be a lot more resilient as unit tests
instead of MTR tests, but that's out-of-scope for this work.)

Change-Id: I9f3728a1f88df55530f85bc79944d18d636ede6c